+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         android:usesCleartextTraffic="true">
-        <activity android:name=".presentation.chatting.ChattingActivity"
+        <activity android:name=".presentation.chatting.view.ChattingActivity"
             android:theme="@style/Theme.MaterialComponents.Light.NoActionBar.Bridge"
             android:windowSoftInputMode="adjustResize"></activity>
         <activity android:name=".presentation.filter.FilterActivity"></activity>
